organized folders and print route
