+++ package.json
@@ -11,6 +11,7 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.4.1",
     "react-scripts": "5.0.1",
+    "react-to-print": "^2.14.7",
     "router": "^1.3.7",
     "styled-components": "^5.3.6",
     "web-vitals": "^2.1.4"
