+++ src/App.js
@@ -2,11 +2,11 @@ import React from 'react';
 import './App.css';
 import { Route, Routes } from 'react-router-dom';
 import { NavBar } from './components/NavBar';
-import { About } from './components/About';
-import { Results } from './components/Results';
+import { About } from './pages/About';
+import { Result } from './components/Result';
 import { Print } from './components/Print';
 import { Charts } from './components/Charts';
-import { Home } from './components/Home';
+import { Home } from './pages/Home';
 
 const App = () => {
   return (
@@ -15,7 +15,7 @@ const App = () => {
        <Routes>
         <Route path='/' element={<Home/>} />
         <Route path='/about' element={<About/>} />
-        <Route path='/result' element={<Results/>} />
+        <Route path='/result' element={<Result/>} />
         <Route path='/result/charts' element={<Charts/>} />
         <Route path='/result/print' element={<Print/>} />
       </Routes>
